refactor(daengle): 컴포넌트명 kebab-case로 변경
