+++ apps/daengle/src/pages/estimate/grooming/[id]/index.tsx
@@ -6,7 +6,7 @@ import { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import dayjs, { Dayjs } from 'dayjs';
 import 'dayjs/locale/ko';
-import EstimateSelectComponent from '~/components/estimate/EstimateSelectComponent';
+import EstimateSelectComponent from '~/components/estimate/estimate-select-component';
 import {
   usePostUserEstimateGroomingMutation,
   usePostUserEstimateGroomerUserInfoMutation,
@@ -14,7 +14,7 @@ import {
 import { PetInfos, PostUserEstimateGroomerUserInfoResponse } from '~/models/estimate';
 import { ROUTES } from '~/constants/commons';
 import { DefaultImage } from '@daengle/design-system/icons';
-import DatePickerComponent from '~/components/estimate/DatePickerComponent';
+import DatePickerComponent from '~/components/estimate/date-picker-component';
 
 export default function EstimateCreate() {
   const router = useRouter();
